Actually fixed it
Thought in the last commit I had fixed the exception handling but I only dug myself a hole. Now it works completely by checking to see if the datatype of get_info is a pandas df or a boolean. I have also added an option for the user to display out all the listed nasdaq tickers to choose from.